fix: refactored FlightTable into smaller components for better readability
